crypto: Factor test material
